init: Fix riscv32-unknown-gcc link
Additionally, check if wget successfully downloaded.

+++ .init.sh
@@ -16,11 +16,19 @@ pip3 install -r requirements.txt
 west init -l zephyr
 
 wget https://github.com/zephyrproject-rtos/sdk-ng/releases/download/v${ZEPHYR_SDK_VERSION}/zephyr-toolchain-riscv64-${ZEPHYR_SDK_VERSION}-linux-x86_64-setup.run
+if ! test -f zephyr-toolchain-riscv64-${ZEPHYR_SDK_VERSION}-linux-x86_64-setup.run; then
+	echo "wget error: Unable to fetch SDK"
+	exit 2
+fi
 chmod +x zephyr-toolchain-riscv64-${ZEPHYR_SDK_VERSION}-linux-x86_64-setup.run
 ./zephyr-toolchain-riscv64-${ZEPHYR_SDK_VERSION}-linux-x86_64-setup.run -- -d $PWD/zephyr-sdk-${ZEPHYR_SDK_VERSION} -y -nocmake
 rm zephyr-toolchain-riscv64-${ZEPHYR_SDK_VERSION}-linux-x86_64-setup.run
 
-wget https://github.com/stnolting/riscv-gcc-prebuilt/releases/download/rv32i-1.0.0/riscv32-unknown-elf.gcc-10.2.0.rv32i.ilp32.newlib.tar.gz
+wget https://github.com/stnolting/riscv-gcc-prebuilt/releases/download/rv32i-2.0.0/riscv32-unknown-elf.gcc-10.2.0.rv32i.ilp32.newlib.tar.gz
+if ! test -f riscv32-unknown-elf.gcc-10.2.0.rv32i.ilp32.newlib.tar.gz; then
+	echo "wget error: Unable to fetch riscv32-unknown-gcc"
+	exit 2
+fi
 mkdir riscv32-unknown-elf
 tar -xvf riscv32-unknown-elf.gcc-10.2.0.rv32i.ilp32.newlib.tar.gz -C riscv32-unknown-elf/
 rm riscv32-unknown-elf.gcc-10.2.0.rv32i.ilp32.newlib.tar.gz
